feat: Implementado acl e login

+++ config/routes.php
@@ -9,7 +9,6 @@ declare(strict_types=1);
  * @contact  group@hyperf.io
  * @license  https://github.com/hyperf/hyperf/blob/master/LICENSE
  */
-
 use App\Constants\Permission;
 use App\Controller\PessoaController;
 use App\Controller\PessoaEnderecoController;
@@ -19,33 +18,35 @@ use Hyperf\HttpServer\Router\Router;
 
 Router::addRoute(['GET', 'POST', 'HEAD'], '/', 'App\Controller\IndexController@index');
 
-// CRUD Pessoa
-Router::get('/pessoa', [PessoaController::class, 'index'],[
-    'middleware' => [AclMiddleware::class],
-    'options' => [
-        'acl_resource' => 'Pessoa.Gerenciamento',
-        'acl_level' => Permission::ACESSAR
-    ]
-]);
+/** @return array<string, mixed> */
+$acl = static function (string $resource, int $level): array {
+    return [
+        'middleware' => [AclMiddleware::class],
+        'acl_resource' => $resource,
+        'acl_level' => $level,
+    ];
+};
 
-Router::get('/pessoa/{id:\d+}', [PessoaController::class, 'show']);
-Router::post('/pessoa', [PessoaController::class, 'store']);
-Router::put('/pessoa/{id:\d+}', [PessoaController::class, 'update']);
-Router::delete('/pessoa/{id:\d+}', [PessoaController::class, 'destroy']);
+// CRUD Pessoa
+Router::get('/pessoa', [PessoaController::class, 'index'], $acl('Pessoa.Gerenciamento', Permission::ACESSAR));
+Router::get('/pessoa/{id:\d+}', [PessoaController::class, 'show'], $acl('Pessoa.Gerenciamento', Permission::ACESSAR));
+Router::post('/pessoa', [PessoaController::class, 'store'], $acl('Pessoa.Gerenciamento', Permission::INSERIR));
+Router::put('/pessoa/{id:\d+}', [PessoaController::class, 'update'], $acl('Pessoa.Gerenciamento', Permission::ALTERAR));
+Router::delete('/pessoa/{id:\d+}', [PessoaController::class, 'destroy'], $acl('Pessoa.Gerenciamento', Permission::REMOVER));
 
 // CRUD PessoaFisica
-Router::get('/pessoa-fisica', [PessoaFisicaController::class, 'index']);
-Router::get('/pessoa-fisica/{id:\d+}', [PessoaFisicaController::class, 'show']);
-Router::post('/pessoa-fisica', [PessoaFisicaController::class, 'store']);
-Router::put('/pessoa-fisica/{id:\d+}', [PessoaFisicaController::class, 'update']);
-Router::delete('/pessoa-fisica/{id:\d+}', [PessoaFisicaController::class, 'destroy']);
+Router::get('/pessoa-fisica', [PessoaFisicaController::class, 'index'], $acl('PessoaFisica.Gerenciamento', Permission::ACESSAR));
+Router::get('/pessoa-fisica/{id:\d+}', [PessoaFisicaController::class, 'show'], $acl('PessoaFisica.Gerenciamento', Permission::ACESSAR));
+Router::post('/pessoa-fisica', [PessoaFisicaController::class, 'store'], $acl('PessoaFisica.Gerenciamento', Permission::INSERIR));
+Router::put('/pessoa-fisica/{id:\d+}', [PessoaFisicaController::class, 'update'], $acl('PessoaFisica.Gerenciamento', Permission::ALTERAR));
+Router::delete('/pessoa-fisica/{id:\d+}', [PessoaFisicaController::class, 'destroy'], $acl('PessoaFisica.Gerenciamento', Permission::REMOVER));
 
 // CRUD PessoaEndereco
-Router::get('/pessoa-endereco', [PessoaEnderecoController::class, 'index']);
-Router::get('/pessoa-endereco/{id:\d+}', [PessoaEnderecoController::class, 'show']);
-Router::post('/pessoa-endereco', [PessoaEnderecoController::class, 'store']);
-Router::put('/pessoa-endereco/{id:\d+}', [PessoaEnderecoController::class, 'update']);
-Router::delete('/pessoa-endereco/{id:\d+}', [PessoaEnderecoController::class, 'destroy']);
+Router::get('/pessoa-endereco', [PessoaEnderecoController::class, 'index'], $acl('PessoaEndereco.Gerenciamento', Permission::ACESSAR));
+Router::get('/pessoa-endereco/{id:\d+}', [PessoaEnderecoController::class, 'show'], $acl('PessoaEndereco.Gerenciamento', Permission::ACESSAR));
+Router::post('/pessoa-endereco', [PessoaEnderecoController::class, 'store'], $acl('PessoaEndereco.Gerenciamento', Permission::INSERIR));
+Router::put('/pessoa-endereco/{id:\d+}', [PessoaEnderecoController::class, 'update'], $acl('PessoaEndereco.Gerenciamento', Permission::ALTERAR));
+Router::delete('/pessoa-endereco/{id:\d+}', [PessoaEnderecoController::class, 'destroy'], $acl('PessoaEndereco.Gerenciamento', Permission::REMOVER));
 
 Router::get('/favicon.ico', function () {
     return '';
